Remove focus event from main body
Currently this focus event shifts the entire window scroll position to sit at the top of the spreadsheets position. Removing this focus event has no other effect that I have seen. If there is a reason for this focus event, can there please be a solution that doesn't involve focusing the main span into view and constantly shifting the windows scroll.

Also, in the scroll shifting, any horizontal scrolling that the spreadsheet sits in, gets reset back to 0. This is a big and annoying issue that this PR resolves